Authentications added for Places controller

+++ public_html/protected/controllers/PlacesController.php
@@ -2,11 +2,23 @@
 class PlacesController extends RestController
 {
     public function actionIndex() {
-        $this->render('index');
+        if (Yii::app()->user->checkAccess('viewPlaces'))
+        {
+            $this->render('index');
+        }
+        else {
+            $this->redirect(array('site/login'));
+        }
     }
 
     public function actionList()
     {
+        if (!Yii::app()->user->checkAccess('viewPlaces'))
+        {
+            $this->_sendResponse(403);
+            return;
+        }
+        //searching only for current users places (defaultScope returns appropriate condition)
         $places = Places::model()->findAll();
         echo CJSON::encode($places);
     }
@@ -18,13 +30,17 @@ class PlacesController extends RestController
 
     public function actionCreate()
     {
+        if (!Yii::app()->user->checkAccess('addPlace'))
+        {
+            $this->_sendResponse(403);
+            return;
+        }
         $data = CJSON::decode(file_get_contents('php://input'));
         $place = new Places();
         $place->p_title = $data['p_title'];
         $place->p_description = isset($data['p_description']) ? $data['p_description'] : '';
         $place->p_lng = $data['p_lng'];
         $place->p_lat = $data['p_lat'];
-        $place->p_user_id = 1;
         if ($place->save()) {
             $this->_sendResponse(200, CJSON::encode($place));
         }
@@ -40,6 +56,11 @@ class PlacesController extends RestController
     {
         $data = CJSON::decode(file_get_contents('php://input'));
         $place = Places::model()->findByPk($id);
+        if (!Yii::app()->user->checkAccess('updatePlace', array('place'=>$place)))
+        {
+            $this->_sendResponse(403);
+            return;
+        }
         if (null === $place) {
             $this->_sendResponse(404, CJSON::encode(array('message'=>'Could not find place with id = '.$id)));
         }
@@ -47,7 +68,6 @@ class PlacesController extends RestController
         $place->p_description = isset($data['p_description']) ? $data['p_description'] : '';
         $place->p_lng = $data['p_lng'];
         $place->p_lat = $data['p_lat'];
-        $place->p_user_id = 1;
         if ($place->save()) {
             $this->_sendResponse(200, CJSON::encode($place));
         }
@@ -63,7 +83,13 @@ class PlacesController extends RestController
     {
         $place = Places::model()->findByPk($id);
         if (null === $place) {
-            $this->_sendResponse(404, CJSON::encode(array('message'=>'Could not delete place with id = '.$id)));
+            $this->_sendResponse(404, CJSON::encode(array('message'=>'Could not find place with id = '.$id)));
+            return;
+        }
+        if (!Yii::app()->user->checkAccess('updatePlace', array('place'=>$place)))
+        {
+            $this->_sendResponse(403);
+            return;
         }
         if ($place->delete()) {
             $this->_sendResponse(200, CJSON::encode($place));
